Revert "Moved ColorSpace and primitives to shared"
