+++ src/core/ccitt_stream.js
@@ -13,7 +13,7 @@
  * limitations under the License.
  */
 
-import { Dict, isDict } from '../shared/primitives';
+import { Dict, isDict } from './primitives';
 import { CCITTFaxDecoder } from './ccitt';
 import { DecodeStream } from './stream';
