+++ src/core/cmap.js
@@ -16,7 +16,7 @@
 import {
   CMapCompressionType, FormatError, isString, unreachable, warn
 } from '../shared/util';
-import { isCmd, isEOF, isName, isStream } from '../shared/primitives';
+import { isCmd, isEOF, isName, isStream } from './primitives';
 import { Lexer } from './parser';
 import { MissingDataException } from './core_utils';
 import { Stream } from './stream';
