+++ src/core/image.js
@@ -13,11 +13,9 @@
  * limitations under the License.
  */
 
-import {
-  assert, ColorSpace, FormatError,
-  ImageKind, info, warn
-} from '../shared/util';
-import { isName, isStream, Name } from '../shared/primitives';
+import { assert, FormatError, ImageKind, info, warn } from '../shared/util';
+import { isName, isStream, Name } from './primitives';
+import { ColorSpace } from './colorspace';
 import { DecodeStream } from './stream';
 import { JpegStream } from './jpeg_stream';
 import { JpxImage } from './jpx';
