+++ src/core/obj.js
@@ -14,21 +14,22 @@
  */
 
 import {
-  assert, bytesToString, ColorSpace, createPromiseCapability,
-  createValidAbsoluteUrl, FormatError, info, InvalidPDFException,
-  isBool, isNum, isString, PermissionFlag, shadow,
-  stringToPDFString, stringToUTF8String, unreachable, warn
+  assert, bytesToString, createPromiseCapability, createValidAbsoluteUrl,
+  FormatError, info, InvalidPDFException, isBool, isNum, isString,
+  PermissionFlag, shadow, stringToPDFString, stringToUTF8String, unreachable,
+  warn
 } from '../shared/util';
 import {
   clearPrimitiveCaches, Cmd, Dict, isCmd, isDict, isName, isRef, isRefsEqual,
   isStream, Ref, RefSet, RefSetCache
-} from '../shared/primitives';
+} from './primitives';
 import { Lexer, Parser } from './parser';
 import {
   MissingDataException, toRomanNumerals, XRefEntryException, XRefParseException
 } from './core_utils';
 import { ChunkedStream } from './chunked_stream';
 import { CipherTransformFactory } from './crypto';
+import { ColorSpace } from './colorspace';
 
 function fetchDestination(dest) {
   return isDict(dest) ? dest.get('D') : dest;
