+++ src/display/canvas.js
@@ -14,10 +14,12 @@
  */
 
 import {
-  ColorSpace, FONT_IDENTITY_MATRIX, IDENTITY_MATRIX,
-  ImageKind, info, isLittleEndian, isNum,
+  FONT_IDENTITY_MATRIX, IDENTITY_MATRIX, ImageKind, info, isLittleEndian, isNum,
   OPS, shadow, TextRenderingMode, unreachable, Util, warn
 } from '../shared/util';
+import {
+  ColorSpace
+} from '../core/colorspace'
 
 import { getShadingPatternFromIR, TilingPattern } from './pattern_helper';
 
@@ -1701,16 +1703,12 @@ var CanvasGraphics = (function CanvasGraphicsClosure() {
       this.current.patternFill = false;
     },
     setStrokeCMYKColor: function CanvasGraphics_setStrokeCMYKColor(c, m, y, k) {
-      let convStrokeArgs = ColorSpace.singletons.cmyk.getRgb([c, m, y, k], 0);
-      this.setStrokeRGBColor(
-        convStrokeArgs[0],
-        convStrokeArgs[1],
-        convStrokeArgs[2]
-      );
+      let convertedCMYKStrokeArgsToRgb = ColorSpace.singletons.cmyk.getRgb([c, m, y, k], 0);
+      this.setStrokeRGBColor(convertedCMYKStrokeArgsToRgb[0], convertedCMYKStrokeArgsToRgb[1], convertedCMYKStrokeArgsToRgb[2])
     },
     setFillCMYKColor: function CanvasGraphics_setFillCMYKColor(c, m, y, k) {
-      let confFillArgs = ColorSpace.singletons.cmyk.getRgb([c, m, y, k], 0);
-      this.setFillRGBColor(confFillArgs[0], confFillArgs[1], confFillArgs[2]);
+      let convertedCMYKFillArgsToRgb = ColorSpace.singletons.cmyk.getRgb([c, m, y, k], 0);
+      this.setFillRGBColor(convertedCMYKFillArgsToRgb[0], convertedCMYKFillArgsToRgb[1], convertedCMYKFillArgsToRgb[2]);
     },
 
     shadingFill: function CanvasGraphics_shadingFill(patternIR) {
