+++ src/shared/util.js
@@ -15,7 +15,6 @@
 /* eslint no-var: error */
 
 import './compatibility';
-import { ColorSpace } from './colorspace';
 import { ReadableStream } from './streams_polyfill';
 
 const IDENTITY_MATRIX = [1, 0, 0, 1, 0, 0];
@@ -888,7 +887,6 @@ export {
   FontType,
   ImageKind,
   CMapCompressionType,
-  ColorSpace,
   AbortException,
   InvalidPDFException,
   MissingPDFException,
