+++ test/unit/colorspace_spec.js
@@ -13,9 +13,9 @@
  * limitations under the License.
  */
 
-import { Dict, Name, Ref } from '../../src/shared/primitives';
+import { Dict, Name, Ref } from '../../src/core/primitives';
 import { Stream, StringStream } from '../../src/core/stream';
-import { ColorSpace } from '../../src/shared/colorspace';
+import { ColorSpace } from '../../src/core/colorspace';
 import { PDFFunctionFactory } from '../../src/core/function';
 import { XRefMock } from './test_utils';
