+++ test/unit/crypto_spec.js
@@ -17,7 +17,7 @@ import {
   AES128Cipher, AES256Cipher, ARCFourCipher, calculateMD5, calculateSHA256,
   calculateSHA384, calculateSHA512, CipherTransformFactory, PDF17, PDF20
 } from '../../src/core/crypto';
-import { Dict, Name } from '../../src/shared/primitives';
+import { Dict, Name } from '../../src/core/primitives';
 import {
   PasswordException, PasswordResponses, stringToBytes
 } from '../../src/shared/util';
