+++ test/unit/test_utils.js
@@ -15,7 +15,7 @@
 
 import { assert, CMapCompressionType } from '../../src/shared/util';
 import isNodeJS from '../../src/shared/is_node';
-import { isRef } from '../../src/shared/primitives';
+import { isRef } from '../../src/core/primitives';
 import { Page } from '../../src/core/document';
 
 class DOMFileReaderFactory {
